Refactoring: apply extract method for readability

+++ src/main/java/it/vallara/saletaxesproblem/TaxRateCondition.java
@@ -1,11 +1,13 @@
 package it.vallara.saletaxesproblem;
 
+import java.math.BigDecimal;
+
 public class TaxRateCondition {
 
     public Condition condition;
-    public double taxRate;
+    public BigDecimal taxRate;
 
-    public TaxRateCondition(Condition condition, double taxRate) {
+    public TaxRateCondition(Condition condition, BigDecimal taxRate) {
         this.condition = condition;
         this.taxRate = taxRate;
     }
